Moved entities from GameVar to Scene
