+++ src/bomberman/GameLoop.java
@@ -1,7 +1,7 @@
 package bomberman;
 
 import bomberman.entity.Entity;
-import bomberman.gamecontroller.GameVariables;
+import bomberman.scenes.Sandbox;
 import javafx.animation.AnimationTimer;
 import javafx.scene.canvas.GraphicsContext;
 
@@ -21,7 +21,8 @@ public class GameLoop
                 tickDuration = (currentNanoTime - startNanoTime) / 1000000000.0;
                 gc.clearRect(0, 0, 512, 512);
 
-                for(Entity e : GameVariables.getEntities())
+                //TODO This will have to be something like, currentScene.getEntities()
+                for(Entity e : Sandbox.getEntities())
                 	e.draw();
             }
         }.start();
