+++ src/bomberman/gamecontroller/GameVariables.java
@@ -5,27 +5,10 @@
  */
 package bomberman.gamecontroller;
 
-import java.util.Vector;
-
-import bomberman.entity.Entity;
-
 /**
  *
  * @author ashish.padalkar
  */
 public class GameVariables {
-	private static Vector<Entity> entities = new Vector<Entity>();
 
-	public static Vector<Entity> getEntities(){
-		return entities;
-	}
-
-	public static boolean addEntityToGame(Entity e){
-		if(!entities.contains(e)){
-			entities.add(e);
-			return true;
-		} else {
-			return false;
-		}
-	}
 }
